requirements.txt
requests
InquirerPy
fpdf2
